% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignClones_ref.R
\name{AssignClones_ref}
\alias{AssignClones_ref}
\title{Using marker regions to assign each cell into c reference subclones}
\usage{
AssignClones_ref(
  rhohats,
  thetahats,
  snpCoverages,
  priorCloneProbs = NULL,
  clone.genotypes,
  sigma.rho
)
}
\arguments{
\item{rhohats:}{n by m matrix of rho_hat values for each cell across the regions.}

\item{thetahats:}{n by m matrix of theta_hat values for each cell across the regions.}

\item{snpCoverages:}{n by m matrix of total read counts covering SNPs for each cell across the regions.}

\item{priorCloneProbs:}{A numeric vector indicating prior prior probability of each subclone.}

\item{clone.genotypes:}{c by m matrix of numbers representing different genotypes for each clone and each maker region (known from scDNA-seq).}

\item{sigma.rho:}{Numeric. Standard deviation of the rho_i values under normal distribution.}
}
\description{
rhohats, thetahats, snpCoverages are n by m matrices for n cell and m marker regions.
The genotypes (rho, theta) are: 1.(0.5,0); 2.(0.5,1); 3.(1,0); 4.(1,0.5); 5.(1,1); 6.(1.5,0); 7.(1.5,1/3); 8.(1.5,2/3); 9.(1.5,1); 10.(2,0); 11.(2,1/4); 12.(2,2/4); 13.(2,3/4); 14.(2,4/4)
15.(2.5,0); 16.(2.5,1/5); 17.(2.5,2/5); 18.(2.5,3/5); 19.(2.5,4/5); 20.(2.5,1); 21.(3,0); 22.(3,1/6); 23.(3,2/6); 24.(3,3/6); 25.(3,4/6); 26.(3,5/6); 27.(3,6/6)
}
