% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cov_value.R
\name{Cov_value}
\alias{Cov_value}
\title{Normalize coverage using identified/ specified normal cells and one normal region and generate a table with (rho_hat) of each cell for all regions.}
\usage{
Cov_value(
  Obj_filtered = NULL,
  type = "tumor",
  raw_counts = NULL,
  ref_counts = NULL,
  cov_adj = 1,
  qt_filter = FALSE,
  refr = TRUE,
  plot_path = NULL
)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object with segments, specified normal cells and a normal region}

\item{type}{Specify whether the sample is a "tumor" or "cellline". If "type" is a "cellline", param "ref_counts" needs to be specified for normal sample.}

\item{raw_counts}{(required) A large binned coverage matrix (m1 bin by n1 cell) with values being read counts for all chromosomal regions of tumor sample.}

\item{ref_counts}{(required only when type = "cellline") A binned coverage matrix (m2 bin by n2 cell) with values being read counts for all chromosomal regions of normal sample. n2 can be 1 for bulk sample.}

\item{cov_adj}{An integer for coverage adjustment for tumor cells.}

\item{qt_filter}{Logical (TRUE/ FALSE). Whether or not to exclude cells with rho_hat>0.99 or <0.01 for each region.}

\item{refr}{Logical (TRUE/ FALSE). Whether or not to use diplid region for normalization (otherwise, cell size is used).}
}
\value{
(rho_hat) of each cell for all region in the "cov_values".
Every column in the cov_values is (rho_hat) of each region. Each row is a cell.
}
\description{
rho_hat: Relative coverage change for each cell in a region
}
