% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_util.R
\name{create_long_data_frame}
\alias{create_long_data_frame}
\title{Creat long dataframe with allele specific states and rho and theta values}
\usage{
create_long_data_frame(Obj_filtered)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object}
}
\value{
dataframe with allele specific states and rho and theta values extracted from the alleloscope object
}
\description{
Creat long dataframe with allele specific states and rho and theta values
}
