% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{Iterative phasing and theta_hat estimation}
\usage{
EM(ref_table, alt_table, max_iter = max_iter, sub_cells = NULL, seed = 2020)
}
\arguments{
\item{ref_table}{A SNP by cell read count matrix/ spare matrix for the reference alleles.}

\item{alt_table}{A SNP by cell read count matrix/ spare matrix for the alternative alleles.}

\item{max_iter}{An integer of maximum iteration number.}

\item{sub_cells}{A vector of cell names for the cells used to estimate the phases.}

\item{seed}{An integer of random seed number for EM initialization.}
}
\value{
A list of estimated indicators (I_hat) for each SNP and estimated major haplotype proportion (theta_hat) for each cell in one region. I_hat is the phasing result indicating whether reference allele is on the major haplotype for each SNP. Theta_hat represents the CNV states for each cell. A cell is considered as a CNV carrier if its theta_hat depart from 0.5.
}
\description{
Iterative phasing and theta_hat estimation
}
