% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_regions.R
\name{Est_regions}
\alias{Est_regions}
\title{Perform EM iterations on the filtered cells with barcodes, and plot the results for each region.}
\usage{
Est_regions(
  Obj_filtered = NULL,
  max_nSNP = 30000,
  plot_stat = TRUE,
  min_cell = 20,
  min_snp = 0,
  rds_path = NULL,
  cont = FALSE,
  max_iter = 50,
  phases = NULL,
  sub_cells = NULL,
  sub_region = NULL
)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object with allele and segment information for estimating cell major haplotype proportion (theta_hat) for each region.}

\item{max_nSNP}{Integer. Maximum SNP number used for estimating theta_hat for a region.}

\item{plot_stat}{Logical (TRUE/ FALSE). Whether or not to plot the statistics and EM results for each region.}

\item{min_cell}{Integer. Filter out the cells with reads < min_cells.}

\item{min_snp}{Integer. Filter out the SNPs with reads < min_snp.}

\item{rds_path}{The path for saving the rds files for the estimated results for each region.}

\item{cont}{Logical (TRUE/FALSE). Whether or not to skip the regions with the rds files in the specified rds_path.}

\item{max_iter}{Integer. Maximum numbers of EM iterations.}

\item{phases}{List. The estimated phase indicators (I_j) of each SNP across all regions.}

\item{sub_cells}{A vector of cell names for the cells used to update the phases.}

\item{sub_region}{A region name (in the "chrr" column of the seg_table) for a region where the SNP phases and cell major haplotype proportion are updated.}
}
\value{
A "rds_list" of the estimated SNP phases (I_hat), estimated cell major haplotype proportion (theta_hat) for all regions.
}
\description{
Perform EM iterations on the filtered cells with barcodes, and plot the results for each region.
}
