% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype.R
\name{Genotype}
\alias{Genotype}
\title{Genotype each cell for each region and plot the genotypes.}
\usage{
Genotype(Obj_filtered = NULL, xmax = NULL, plot_path = NULL, ref_gt = NULL)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object with a n cell by (m region * 4) genotype_values matrix and seg_table_filtered matrix.
Every 2 columns in the genotype_table matrix are (rho_hat, theta_hat) of each region.}

\item{xmax}{An integer for the x-axis maximum limit.}

\item{plot_path}{The path for saving the plot.}

\item{ref_gt}{A reference "genotypes" (from scDNA-seq) to help with genotype estimation.}
}
\value{
A list of ggplot objects of the genotyping results for all the regions.
}
\description{
Genotype each cell for each region and plot the genotypes.
}
