% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_conf.R
\name{genotype_conf}
\alias{genotype_conf}
\title{Compute confidence scores based on posterior probability for each cell in a region.}
\usage{
genotype_conf(X = NULL, gt = NULL)
}
\arguments{
\item{X:}{A ncell by 2 dataframe. Column 1: normalized coverage (rho_hat); Column 2: theta_hat}

\item{gt:}{A vector of lenth ncell. The numbers represent cell-level allele-specific copy number states.}
}
\value{
A lineage tree plot constructed using cell-level haplotype profiles across all regions.
}
\description{
Compute confidence scores based on posterior probability for each cell in a region.
}
