% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_ref.R
\name{genotype_ref}
\alias{genotype_ref}
\title{Genotype each cell based on coverage and major haplotype proportion (theta_hat) in a region.}
\usage{
genotype_ref(X, snpCoverage, possibleGenotypes, priorProb, sigma.rho = 0.25)
}
\arguments{
\item{X:}{A ncell by 2 dataframe. Column 1: normalized coverage (rho_hat); Column 2: theta_hat}

\item{snpCoverage:}{A numeric vector. Total read counts covering SNPs for each cell in a region.}

\item{possibleGenotypes:}{A numeric vector indicating all possible genotypes in a region.}

\item{priorProb:}{A numeric vector indicating prior prior probability of each possible genotype. The order should match the one for "possibleGenotypes".}

\item{sigma.rho:}{Numeric. Standard deviation of the rho_i values under normal distribution.}
}
\description{
possibleGenotypes, priorProb come from referece data.
}
\keyword{internal}
