% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_value.R
\name{Genotype_value}
\alias{Genotype_value}
\title{Normalize coverage using identified/ specified normal cells and one normal region and generate a table with (rho_hat, theta_hat) of each cell for all regions.}
\usage{
Genotype_value(
  Obj_filtered = NULL,
  type = "tumor",
  raw_counts = NULL,
  ref_counts = NULL,
  cov_adj = 1,
  ref_gtv = NULL,
  mincell = 100,
  qt_filter = TRUE,
  cell_filter = TRUE
)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object with theta_hat info in the rds_list and identified/ specified normal cells and a normal region}

\item{type}{Specify whether the sample is a "tumor" or "cellline". If "type" is a "cellline", param "ref_counts" needs to be specified for normal sample.}

\item{raw_counts}{(required) A large binned coverage matrix (m1 bin by n1 cell) with values being read counts for all chromosomal regions of tumor sample.}

\item{ref_counts}{(required only when type = "cellline") A binned coverage matrix (m2 bin by n2 cell) with values being read counts for all chromosomal regions of normal sample. n2 can be 1 for bulk sample.}

\item{cov_adj}{An integer for coverage adjustment for tumor cells.}

\item{ref_gtv}{A reference "genotype_values" (from scDNA-seq) to help with rho_i estimation.}

\item{mincell}{An integer to filter out regions with minimum number of cells.}

\item{cell_filter}{Logical (TRUE/ FALSE). Whether or not to exclude low quality cells in the output matrix.}
}
\value{
(rho_hat, theta_hat) of each cell for all region in the "genotype_values".
Every 2 columns in the genotype_table are (rho_hat, theta_hat) of each region. Each row is a cell.
}
\description{
rho_hat: Relative coverage change for each cell in a region
theta_hat: Major haplotype proportion fir each cell in a region
}
