% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lineage_plot.R
\name{Lineage_plot}
\alias{Lineage_plot}
\title{Generate genotype plot (scatter plot) for each region and save in the plot directory.}
\usage{
Lineage_plot(
  Obj_filtered = NULL,
  nSNP = 2000,
  clust_method = "ward.D2",
  nclust = 5,
  plot_path = NULL
)
}
\arguments{
\item{Obj_filtered}{A Alleloscope object with a n cell by (m region * 2) genotype_values matrix and seg_table_filtered matrix.
Every 2 columns in the genotype_values matrix are (rho_hat, theta_hat) of each region.}

\item{nSNP}{An integer for the minimum number of SNPs across segments. Segments with the number of SNPs < nSNP are excluded.}

\item{clust_method}{Method for clustering. Please refer to the "pheatmap" function.}

\item{nclust}{An integer for the number of subclones gapped in the plot.}

\item{plot_path}{The path for saving the plot.}
}
\value{
A lineage tree plot constructed using cell-level genotypes across all regions.
}
\description{
Generate genotype plot (scatter plot) for each region and save in the plot directory.
}
