% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_util.R
\name{map_genotypes}
\alias{map_genotypes}
\title{Map genotypes to allele specific states}
\usage{
map_genotypes(Obj_filtered)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object with a n cell by (m region * 4) genotype_values matrix.
Every 2 columns in the genotype_table matrix are (rho_hat, theta_hat) of each region.}
}
\value{
An alleloscope object with an additional slot "genotype_states" that has the from M|m where M is the
major allele copy number and m is the minor allele copy number
}
\description{
Map genotypes to allele specific states
}
