% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_filter.R
\name{Matrix_filter}
\alias{Matrix_filter}
\title{Filter object based on cell number for each SNP, SNP number for each cell, SNP variant allele frequency, and exclude the centromere and telomere regions.}
\usage{
Matrix_filter(
  Obj = NULL,
  cell_filter = 5,
  SNP_filter = 10,
  min_vaf = 0,
  max_vaf = 1,
  centro = NULL,
  telo = NULL,
  plot_stat = TRUE,
  plot_vaf = TRUE
)
}
\arguments{
\item{Obj}{A Alleloscope object.}

\item{cell_filter}{An integer of minimum cell number for SNP selection.}

\item{SNP_filter}{An integer of minimum SNP number for cell selection.}

\item{min_vaf}{A numerical value in the range (0,1) of minimum SNP variant allele frequency in the pseudo bulk for SNP selection.}

\item{max_vaf}{A numerical value in the range (0,1) of mzsimum SNP variant allele frequency in the pseudo bulk for SNP selection.}

\item{centro}{A Matrix/ data.frame of centromere information.}

\item{telo}{A Matrix/ data.frame of telomere information.}

\item{plot_vaf}{Logical (TRUE/FALSE). Whether or not to plot the variant allele frequency for the pseudo bulk for all the chromosomes.}
}
\value{
A Alleloscope object after the filtering.
}
\description{
Filter object based on cell number for each SNP, SNP number for each cell, SNP variant allele frequency, and exclude the centromere and telomere regions.
}
