% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rundf_dna.R
\name{Rundf_dna}
\alias{Rundf_dna}
\title{Run all steps for scDNA-seq data}
\usage{
Rundf_dna(
  alt_all = NULL,
  ref_all = NULL,
  var_all = NULL,
  samplename = "sample",
  genome_assembly = "GRCh38",
  dir_path = "./",
  barcodes = NULL,
  size = NULL,
  assay = "scDNAseq",
  raw_counts = NULL,
  ref_counts = NULL,
  type = "tumor",
  cell_filter = 5,
  SNP_filter = 10,
  min_vaf = 0,
  max_vaf = 1
)
}
\arguments{
\item{alt_all}{A SNP by cell read count matrix/ spare matrix for the alternative alleles.}

\item{ref_all}{A SNP by cell read count matrix/ spare matrix for the reference alleles.}

\item{samplename}{Sample name for the data.}

\item{genome_assembly}{The genome assembly used for sequencing alignment. (ex: "GRCh38" or "GRCh37")}

\item{dir_path}{Path of the output directory.}

\item{barcodes}{A matrix/ data.frame with barcodes for each cell in the first column.}

\item{size}{A numeric vector for the size (bp) of different chromosomes (with the names indicating which chromosome from 1 to 22)}

\item{assay}{A character indicating the type of sequencing data. (ex: "scDNAseq" or "scATACseq")}

\item{raw_counts}{A large binned coverage matrix (m1 bin by n1 cell) for all chromosomal regions of tumor sample.}

\item{ref_counts}{A large binned coverage matrix (m2 bin by n2 cell) for all chromosomal regions of normal sample.}

\item{type}{Specify whethere the sample is a "tumor" or "cellline". If "type" is a "cellline", param "ref_counts" needs to be specified for normal sample.}

\item{cell_filter}{An integer of minimum cell number for SNP selection.}

\item{SNP_filter}{An integer of minimum SNP number for cell selection.}

\item{min_vaf}{A numerical value in the range (0,1) of minimum SNP variant allele frequency in the pseudo bulk for SNP selection.}

\item{max_vaf}{A numerical value in the range (0,1) of mzsimum SNP variant allele frequency in the pseudo bulk for SNP selection.}

\item{vcf_all}{A matrix/ data.frame of the vcf format for SNP information. (The length and order are the same as nrow(alt_all) and nrow(alt_all))}
}
\value{
A Alleloscope object including the necessary information.
}
\description{
Run all steps for scDNA-seq data
}
