% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Createobj.R
\name{Createobj}
\alias{Createobj}
\title{Generate Alleloscope object for analysis}
\usage{
Createobj(
  alt_all = NULL,
  ref_all = NULL,
  var_all = NULL,
  samplename = "sample",
  genome_assembly = "GRCh38",
  dir_path = "./",
  barcodes = NULL,
  size = NULL,
  assay = "scDNAseq",
  cov = FALSE
)
}
\arguments{
\item{alt_all}{A SNP by cell read count matrix/ spare matrix for the alternative alleles.}

\item{ref_all}{A SNP by cell read count matrix/ spare matrix for the reference alleles.}

\item{samplename}{Sample name for the data.}

\item{genome_assembly}{The genome assembly used for sequencing alignment. (ex: "GRCh38" or "GRCh37")}

\item{dir_path}{Path of the output directory.}

\item{barcodes}{A matrix/ data.frame with barcodes for each cell in the first column.}

\item{size}{A matrix with two columns: col1: different chromosome; col2: for the size (bp) of different chromosomes.}

\item{assay}{A character indicating the type of sequencing data. (ex: "scDNAseq" or "scATACseq")}

\item{cov}{Logical (TRUE/FALSE). Whether or not to use only coverage. If "cov" is TRUE, alt_all, ref_all, andvcf_all are not required.}

\item{vcf_all}{A matrix/ data.frame of the vcf format for SNP information. (The length and order are the same as nrow(alt_all) and nrow(alt_all))}
}
\value{
A Alleloscope object including the necessary information.
}
\description{
Generate Alleloscope object for analysis
}
