% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{Segmentation}
\alias{Segmentation}
\title{HMM segmentation based on coverage matrix for paired tumor and normal sample.}
\usage{
Segmentation(
  Obj_filtered = NULL,
  raw_counts = NULL,
  ref_counts = NULL,
  hmm_states = c(0.5, 1.5, 1.8),
  hmm_sd = 0.2,
  plot_seg = TRUE,
  rds_path = NULL,
  adj = 0
)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object.}

\item{raw_counts}{A binned coverage matrix (m1 bin by n1 cell) with values being read counts in DNA sequencing data for all chromosomal regions of tumor sample. n1 can be 1 for bulk sample.}

\item{ref_counts}{A binned coverage matrix (m2 bin by n2 cell) with values being read counts in DNA sequencing data for all chromosomal regions of normal sample. n2 can be 1 for bulk sample.
Numbers of bins (rows) should be the same in the paired chromosomal regions for the paired samples}

\item{hmm_states}{An ordered vector for the HMM states (deletion, 1-copy gain, 2-copy gains).}

\item{plot_seg}{Logical (TRUE/ FALSE). Whether or not to plot the segmentation result.}

\item{rds_path}{The path for saving the rds files for the estimated results for each region.}

\item{adj}{Numeric. Value for tumor coverage adjustment.}

\item{hmm_p}{Numeric. Transition probability for the HMM algorithm.}
}
\value{
A Alleloscope object with "seg_table" added.
}
\description{
If there is no paired normal, other normal sample with the same genome coordinate also works.
}
