% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segments_filter.R
\name{Segments_filter}
\alias{Segments_filter}
\title{Select the segments in the "seg_table" with more than nSNP}
\usage{
Segments_filter(Obj_filtered = NULL, nSNP = 2000)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object with SNP info and raw segmentation table "seg_table".}

\item{nSNP}{An integer of minimum number of SNPs for region selecgtion.}
}
\value{
A Alleloscope object with "seg_table_filtered" added.
}
\description{
Select the segments in the "seg_table" with more than nSNP
}
