% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_normal.R
\name{Select_normal}
\alias{Select_normal}
\title{Identify candidate normal cells and normal regions for cell coverage normalization}
\usage{
Select_normal(Obj_filtered = NULL, raw_counts = NULL, cell_nclust = 5)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object with major haplotype proportion (theta_hat) for each cell of each region in the "rds_list".}

\item{raw_counts}{A large binned coverage matrix (bin by cell) with values being read counts for all chromosomal regions of tumor sample.}
}
\value{
A Alleloscope object with a "select_normal" list added.
A "select_normal" list includes
"barcode_normal": Barcodes of the identified normal cells in the tumor sample.
"region_normal": A vector of ordered potential normal regions for selection. (1st is the most possible.)
"region_normal_rank": A table with the potential "normal regions" for the k clusters from hierarchical clustering.
"k_normal": An integer indicates the kth clsuter that is idenfied as "normal cells"
}
\description{
Identify candidate normal cells and normal regions for cell coverage normalization
}
